refactor(agentic): decompose complex methods and optimize code health for CodeScene
