+++ backend/app/services/agentic/schemas.py
@@ -58,6 +58,20 @@ class TrendingScrapeBatch(BaseModel):
     )
 
 
+def _parse_candidate_confidence(raw_conf: Any) -> float:
+    """Parse string or numeric confidence score into normalized float (0.0 - 1.0)."""
+    if isinstance(raw_conf, (int, float)):
+        return float(raw_conf)
+    if isinstance(raw_conf, str):
+        cleaned = raw_conf.replace("%", "").strip()
+        try:
+            val = float(cleaned)
+            return val / 100.0 if val > 1.0 else val
+        except ValueError:
+            pass
+    return 0.85
+
+
 class SelectorCandidate(BaseModel):
     """Candidate CSS or XPath selector generated by vision diagnosis."""
 
@@ -96,13 +110,8 @@ class SelectorCandidate(BaseModel):
             }
         if isinstance(data, dict):
             raw_conf = data.get("confidence") or data.get("score") or data.get("weight")
-            if isinstance(raw_conf, str):
-                cleaned_conf = raw_conf.replace("%", "").strip()
-                try:
-                    val = float(cleaned_conf)
-                    data["confidence"] = val / 100.0 if val > 1.0 else val
-                except ValueError:
-                    data["confidence"] = 0.85
+            if raw_conf is not None:
+                data["confidence"] = _parse_candidate_confidence(raw_conf)
             if not data.get("reasoning") and data.get("description"):
                 data["reasoning"] = str(data["description"])
         return data
