+++ backend/scripts/demo_self_healing_scrape.py
@@ -16,8 +16,8 @@ import asyncio
 import json
 import logging
 import sys
-import time
 from pathlib import Path
+from typing import Any
 
 # Add backend directory to sys.path
 sys.path.append(str(Path(__file__).parent.parent))
@@ -98,20 +98,12 @@ def sub_header(step: str, title: str) -> None:
     print(f"{DIM}{'-' * 70}{RESET}")
 
 
-async def run_detailed_scrape_demo() -> None:
-    banner("LINKX LIVE DEMO: SCRAPER WITH SELF-HEALING SUPERVISOR", CYAN)
-
-    demo_config_path = Path("/tmp/demo_scrape_config.json")
-    mock_html_path = Path("/tmp/mock_x_trending.html")
-
-    # Save mock HTML file
+def _setup_demo_config(demo_config_path: Path, mock_html_path: Path) -> dict[str, Any]:
+    """Setup mock HTML and broken configuration for demonstration."""
     mock_html_path.write_text(MOCK_X_TRENDING_PAGE_HTML, encoding="utf-8")
-
-    # 1. Setup intentionally broken scrape_config.json
     initial_config = {
         "max_topics_to_scrape": 3,
         "selectors": {
-            # BROKEN on purpose to simulate UI selector break on X.com
             "sidebar_container": "div[data-testid='broken_sidebar_container_v99']",
             "sidebar_link": "a[href*='/search?q=']",
             "tweet_container": "[data-testid='tweet']",
@@ -120,134 +112,87 @@ async def run_detailed_scrape_demo() -> None:
             "exclude_texts": ["Show more", "Subscribe"],
         },
     }
-
     with open(demo_config_path, "w", encoding="utf-8") as f:
         json.dump(initial_config, f, indent=2)
+    return initial_config
+
+
+async def _execute_ai_diagnosis(page: Any, broken_sel: str) -> SelectorDiagnosisReport:
+    """Diagnose broken selector and produce candidate selectors via LLM."""
+    dom_snippet = await get_dom_snippet(page=page, max_chars=3000)
+    model = get_chat_model(temperature=0.1)
+    structured_model = model.with_structured_output(
+        SelectorDiagnosisReport, method="json_mode"
+    )
+    prompt = (
+        f"Failed Element Name: selectors.sidebar_container (Trending Topics Sidebar on X.com)\n"
+        f"Broken Selector: {broken_sel}\n"
+        f"Page DOM:\n{dom_snippet}\n\n"
+        f"Diagnose why the element failed and suggest ranked candidate selectors to target the trending sidebar. Return a JSON object with 'candidate_selectors' list."
+    )
+    diagnosis: SelectorDiagnosisReport = await structured_model.ainvoke(
+        [
+            {
+                "role": "system",
+                "content": "You are a senior browser automation diagnostics agent.",
+            },
+            {"role": "user", "content": prompt},
+        ]
+    )
+    return diagnosis
+
+
+async def _probe_candidates(
+    page: Any, diagnosis: SelectorDiagnosisReport
+) -> str | None:
+    """Test candidate selectors on live page and return first verified match."""
+    for candidate in diagnosis.candidate_selectors:
+        test_res = await validate_selector_candidate(
+            page=page, selector=candidate.selector
+        )
+        if test_res["found"] and test_res["visible"]:
+            print(
+                f"      ✅ {GREEN}{BOLD}MATCH FOUND!{RESET} Visible: True, Match Count: {test_res['count']}"
+            )
+            return candidate.selector
+    return None
+
+
+async def run_detailed_scrape_demo() -> None:
+    banner("LINKX LIVE DEMO: SCRAPER WITH SELF-HEALING SUPERVISOR", CYAN)
+    demo_config_path = Path("/tmp/demo_scrape_config.json")
+    mock_html_path = Path("/tmp/mock_x_trending.html")
+    initial_config = _setup_demo_config(demo_config_path, mock_html_path)
 
     sub_header("SETUP", "Configuration & Initial State")
     print(f"📄 Target Configuration File : {BOLD}{demo_config_path}{RESET}")
-    print(
-        f"🔴 Configured Sidebar Selector : {RED}{initial_config['selectors']['sidebar_container']}{RESET} (Broken)"
-    )
-    print(
-        f"🔗 Configured Link Selector    : {GREEN}{initial_config['selectors']['sidebar_link']}{RESET}"
-    )
 
     playwright_args = get_playwright_args()
 
     async with async_playwright() as p:
         sub_header("STEP 1", "Launch Chrome & Initialize Stealth Session")
-        print("🌐 Launching Chrome (channel='chrome', rebrowser-playwright)...")
         browser = await p.chromium.launch(
-            channel="chrome",
-            headless=True,
-            args=playwright_args,
+            channel="chrome", headless=True, args=playwright_args
         )
         page = await browser.new_page()
         await page.goto(f"file://{mock_html_path}", wait_until="domcontentloaded")
-        print(f"✅ Loaded live page DOM from file://{mock_html_path}")
 
-        # Working copy of in-memory dictionary
         selectors_dict = json.loads(json.dumps(initial_config))
-
-        sub_header("STEP 2", "Scraper Invocation & Fast Pre-Flight Check")
         broken_sel = selectors_dict["selectors"]["sidebar_container"]
-        print(f"🔍 Probing configured selector: {YELLOW}'{broken_sel}'{RESET}...")
 
-        probe = await validate_selector_candidate(
-            page=page, selector=broken_sel, timeout_ms=500
-        )
-        if not probe["found"]:
-            print(
-                f"❌ {RED}{BOLD}ELEMENT MISS DETECTED!{RESET} Selector '{broken_sel}' was not found on the page."
-            )
-            print(
-                f"🚨 {YELLOW}Routing control to LangGraph Self-Healing Supervisor...{RESET}"
-            )
-        else:
-            print("✅ Element found immediately.")
-
-        sub_header("STEP 3", "Node 1: DOM Snippet Extraction (capture_dom)")
-        dom_snippet = await get_dom_snippet(page=page, max_chars=3000)
-        print(f"📦 Pruned & sanitized semantic DOM snippet ({len(dom_snippet)} chars):")
-        print(f"{DIM}{dom_snippet[:350]} ... [truncated]{RESET}")
-
-        sub_header("STEP 4", "Node 2: AI Diagnostic Reasoning (diagnose_dom)")
-        print(
-            f"🤖 Invoking {BOLD}gemini-3.7-flash-high{RESET} with {BOLD}SelectorDiagnosisReport{RESET} structured output..."
-        )
-
-        model = get_chat_model(temperature=0.1)
-        structured_model = model.with_structured_output(
-            SelectorDiagnosisReport, method="json_mode"
-        )
-
-        prompt = (
-            f"Failed Element Name: selectors.sidebar_container (Trending Topics Sidebar on X.com)\n"
-            f"Broken Selector: {broken_sel}\n"
-            f"Page DOM:\n{dom_snippet}\n\n"
-            f"Diagnose why the element failed and suggest ranked candidate selectors to target the trending sidebar. Return a JSON object with 'candidate_selectors' list."
-        )
-
-        t0 = time.perf_counter()
-        diagnosis: SelectorDiagnosisReport = await structured_model.ainvoke(
-            [
-                {
-                    "role": "system",
-                    "content": "You are a senior browser automation diagnostics agent. Return a valid SelectorDiagnosisReport.",
-                },
-                {"role": "user", "content": prompt},
-            ]
-        )
-        latency = time.perf_counter() - t0
-
-        print(f"⚡ LLM Response Received in {BOLD}{latency:.2f}s{RESET}:")
-        print(f"   • {BOLD}Broken Element{RESET} : {diagnosis.broken_element_name}")
-        print(
-            f"   • {BOLD}Page State{RESET}     : {GREEN}{diagnosis.page_state}{RESET}"
-        )
-        print(
-            f"   • {BOLD}Recoverable{RESET}    : {GREEN}{diagnosis.is_recoverable}{RESET}"
-        )
-        print(f"\n   🎯 {BOLD}Ranked Candidate Selectors Proposed by AI:{RESET}")
-        for i, c in enumerate(diagnosis.candidate_selectors, 1):
-            print(
-                f"      {CYAN}{i}. {c.selector}{RESET} (Confidence: {GREEN}{c.confidence * 100:.0f}%{RESET})"
-            )
-            print(f"         {DIM}Reasoning: {c.reasoning}{RESET}")
-
-        sub_header(
-            "STEP 5",
-            "Node 3: Live Playwright Page Candidate Probing (verify_candidates)",
-        )
-        verified_selector = None
-        for i, candidate in enumerate(diagnosis.candidate_selectors, 1):
-            print(
-                f"   🧪 Testing Candidate {i}: {CYAN}'{candidate.selector}'{RESET} on live page..."
-            )
-            test_res = await validate_selector_candidate(
-                page=page, selector=candidate.selector
-            )
-            if test_res["found"] and test_res["visible"]:
-                print(
-                    f"      ✅ {GREEN}{BOLD}MATCH FOUND!{RESET} Visible: True, Match Count: {test_res['count']}"
-                )
-                verified_selector = candidate.selector
-                break
-            else:
-                print("      ❌ Candidate failed (Found: False)")
+        sub_header("STEP 2", "AI Diagnostic Reasoning & Verification")
+        diagnosis = await _execute_ai_diagnosis(page, broken_sel)
+        verified_selector = await _probe_candidates(page, diagnosis)
 
         if not verified_selector:
             print(f"❌ {RED}All candidate selectors failed.{RESET}")
             return
 
-        sub_header("STEP 6", "Node 4: Hot-Patching Configuration on Disk (apply_patch)")
         patch_success = patch_selector_config(
             config_path=demo_config_path,
             key_path="selectors.sidebar_container",
             new_selector=verified_selector,
         )
-        # Update in-memory dict
         selectors_dict["selectors"]["sidebar_container"] = verified_selector
 
         print(
